requirements.txt
mysql-connector-python~=8.3.0
requests~=2.31.0
gql~=3.5.0
schedule~=1.2.1
uvicorn~=0.29.0
fastapi~=0.110.2
requests-toolbelt~=1.0.0
XlsxWriter~=3.2.0
starlette~=0.37.2
PyYAML~=6.0.1